refactor(ship): éclate service.ts en dossier services/ + conso en namespace shipService
