+++ apps/bot/src/domains/ship/index.ts
@@ -1,3 +1,2 @@
 export { MAX_SHIP_NAME_LENGTH } from './constants.js';
-export { findOrCreateShip } from './service.js';
 export { shipButtonHandlers } from './interactions/index.js';
