+++ apps/bot/src/domains/ship/interactions/confirm-upgrade-button.ts
@@ -6,7 +6,7 @@ import { buildOpEmbed } from '../../../discord/utils/build-op-embed.js';
 import { assertInteractorIsTheOwner, editReply, parseIntegerArg, parseOwnerDiscordId } from '../../../discord/utils/index.js';
 import { CONFIRM_SHIP_MODULE_UPGRADE_BUTTON_NAME } from '../constants.js';
 import { SHIP_MODULE_LABELS } from '../modules.js';
-import { upgradeShipModule } from '../service.js';
+import * as shipService from '../services/index.js';
 import { parseShipModuleKey } from '../utils/index.js';
 
 async function handle(interaction: ButtonInteraction, args: Array<string>): Promise<void> {
@@ -17,7 +17,7 @@ async function handle(interaction: ButtonInteraction, args: Array<string>): Prom
   const expectedUpgradeLevel = parseIntegerArg(args[3]);
 
   await interaction.deferUpdate();
-  const upgradedShip = await upgradeShipModule(playerId, moduleKey, expectedUpgradeLevel);
+  const upgradedShip = await shipService.upgradeShipModule(playerId, moduleKey, expectedUpgradeLevel);
   const upgradedLevel = upgradedShip[SHIP_MODULE_LEVEL_COLUMNS[moduleKey]];
 
   const embed = buildOpEmbed('success')
