--- apps/bot/src/domains/ship/service.ts
@@ -1,118 +1,0 @@
-import { db, SHIP_MODULE_LEVEL_COLUMNS, type DbOrTransaction, type Ship, type ShipModuleKey } from '@one-piece/db';
-
-import { ValidationError } from '../../discord/errors.js';
-import { sanitizeName } from '../../shared/sanitize-name.js';
-import * as economyRepository from '../economy/repository.js';
-import * as historyRepository from '../history/index.js';
-import * as playerRepository from '../player/repository.js';
-import * as resourceRepository from '../resource/repository.js';
-import { debitResourcesByName } from '../resource/service.js';
-
-import { MAX_SHIP_NAME_LENGTH } from './constants.js';
-import * as shipRepository from './repository.js';
-import { DEFAULT_SHIP_TEMPLATE_KEY, getShipTemplate, type ShipTemplateKey } from './templates.js';
-import type { ShipModuleUpgradePreview, ShipTemplateState } from './types.js';
-import { buildShipModuleUpgradePreview, getMaxHpForHullLevel, getShipModuleBerryCost, getShipModuleResourceCosts } from './utils/index.js';
-
-type FindOrCreateResult = { ship: Ship; created: boolean };
-
-export async function findOrCreateShip(playerId: number, client: DbOrTransaction = db): Promise<FindOrCreateResult> {
-  const existing = await shipRepository.findByPlayerId(playerId, client);
-  if (existing) return { ship: existing, created: false };
-  const created = await shipRepository.create(playerId, buildShipTemplateState(DEFAULT_SHIP_TEMPLATE_KEY), client);
-  return { ship: created, created: true };
-}
-
-/** Change de navire : reset complet aux valeurs de départ du template (les niveaux investis sur l'ancien sont perdus). */
-export async function switchShipTemplate(playerId: number, templateKey: ShipTemplateKey): Promise<Ship> {
-  return db.transaction(async (transaction) => {
-    const current = await shipRepository.findByPlayerIdOrThrow(playerId, transaction, { forUpdate: true });
-    const updatedShip = await shipRepository.resetToTemplate(current.id, buildShipTemplateState(templateKey), transaction);
-
-    await historyRepository.appendHistory({
-      type: 'ship.templateSwitched',
-      payload: { oldTemplate: current.templateKey, newTemplate: templateKey },
-      actorPlayerId: playerId,
-      target: { type: 'ship', id: updatedShip.id },
-      client: transaction,
-    });
-
-    return updatedShip;
-  });
-}
-
-export async function renameShip(playerId: number, newName: string): Promise<Ship> {
-  const sanitized = sanitizeName(newName);
-  if (sanitized.length === 0) {
-    throw new ValidationError('Tu dois donner un nom.');
-  }
-  if (sanitized.length > MAX_SHIP_NAME_LENGTH) {
-    throw new ValidationError(`Le nom du bateau ne peut pas dépasser ${MAX_SHIP_NAME_LENGTH} caractères.`);
-  }
-  return db.transaction(async (transaction) => {
-    const { ship } = await findOrCreateShip(playerId, transaction);
-    const renamed = await shipRepository.rename(ship.id, sanitized, transaction);
-
-    await historyRepository.appendHistory({
-      type: 'ship.renamed',
-      payload: {
-        oldName: ship.name,
-        newName: renamed.name,
-      },
-      actorPlayerId: playerId,
-      target: { type: 'ship', id: renamed.id },
-      client: transaction,
-    });
-
-    return renamed;
-  });
-}
-
-export async function getShipModuleUpgradePreview(playerId: number, moduleKey: ShipModuleKey): Promise<ShipModuleUpgradePreview | null> {
-  const [ship, player, inventory] = await Promise.all([
-    shipRepository.findByPlayerIdOrThrow(playerId),
-    playerRepository.findByIdOrThrow(playerId),
-    resourceRepository.getInventory(playerId),
-  ]);
-
-  return buildShipModuleUpgradePreview({
-    ship,
-    moduleKey,
-    ownedBerries: player.berries,
-    inventory,
-  });
-}
-
-export async function upgradeShipModule(playerId: number, moduleKey: ShipModuleKey, expectedLevel: number): Promise<Ship> {
-  return db.transaction(async (transaction) => {
-    const ship = await shipRepository.findByPlayerIdOrThrow(playerId, transaction, { forUpdate: true });
-    const level = ship[SHIP_MODULE_LEVEL_COLUMNS[moduleKey]];
-    if (level !== expectedLevel) {
-      throw new ValidationError('Cette amélioration a déjà été prise en compte.');
-    }
-
-    const nextLevel = level + 1;
-    const berryCost = getShipModuleBerryCost(moduleKey, level);
-    if (berryCost > 0n) {
-      await economyRepository.debitBerry(playerId, berryCost, transaction);
-    }
-
-    await debitResourcesByName(playerId, getShipModuleResourceCosts(moduleKey, level), transaction);
-
-    return shipRepository.updateModuleLevel(ship.id, moduleKey, nextLevel, transaction);
-  });
-}
-
-function buildShipTemplateState(templateKey: ShipTemplateKey): ShipTemplateState {
-  const { label, startingLevels } = getShipTemplate(templateKey);
-  return {
-    templateKey,
-    name: label,
-    hp: getMaxHpForHullLevel(startingLevels.hull),
-    hullLevel: startingLevels.hull,
-    sailLevel: startingLevels.sail,
-    decksLevel: startingLevels.decks,
-    cabinsLevel: startingLevels.cabins,
-    cargoLevel: startingLevels.cargo,
-  };
-}
