+++ apps/bot/src/domains/ship/services/index.ts
@@ -1,0 +1,5 @@
+export { findOrCreateShip } from './find-or-create-ship.js';
+export { getShipModuleUpgradePreview } from './get-ship-module-upgrade-preview.js';
+export { renameShip } from './rename-ship.js';
+export { switchShipTemplate } from './switch-ship-template.js';
+export { upgradeShipModule } from './upgrade-ship-module.js';
