+++ apps/bot/src/domains/ship/utils/build-ship-template-state.ts
@@ -1,0 +1,18 @@
+import { getShipTemplate, type ShipTemplateKey } from '../templates.js';
+import type { ShipTemplateState } from '../types.js';
+
+import { getMaxHpForHullLevel } from './ship-module-helpers.js';
+
+export function buildShipTemplateState(templateKey: ShipTemplateKey): ShipTemplateState {
+  const { label, startingLevels } = getShipTemplate(templateKey);
+  return {
+    templateKey,
+    name: label,
+    hp: getMaxHpForHullLevel(startingLevels.hull),
+    hullLevel: startingLevels.hull,
+    sailLevel: startingLevels.sail,
+    decksLevel: startingLevels.decks,
+    cabinsLevel: startingLevels.cabins,
+    cargoLevel: startingLevels.cargo,
+  };
+}
