+++ apps/bot/src/domains/ship/utils/index.ts
@@ -1,4 +1,5 @@
 export { buildShipModuleUpgradePreview } from './build-ship-module-upgrade-preview.js';
+export { buildShipTemplateState } from './build-ship-template-state.js';
 export { parseShipModuleKey } from './parse-ship-module-key.js';
 export {
   getMaxHpForHullLevel,
